requirements.txt
fastapi==0.112.2
passlib==1.7.4
pydantic==2.8.2
requests==2.28.2
sqlalchemy==2.0.23
starlette==0.38.2
textstat==0.7.4
pytest==8.3.2
python-dotenv==1.0.1
coverage==7.3.3
httpx==0.27.0
itsdangerous==2.1.2
jinja2==3.1.2
python-multipart==0.0.9
